refactor: rename basic_methods.rs -> commands.rs, format && refactor code
