+++ src/models.rs
@@ -1,20 +1,18 @@
-
 use teloxide::dispatching::dialogue::InMemStorage;
 use teloxide::prelude::Dialogue;
 use teloxide::utils::command::BotCommands;
 
 pub type MyDialogue = Dialogue<State, InMemStorage<State>>;
 
-/// These commands are supported:
 #[derive(BotCommands, Clone)]
 #[command(
     rename_rule = "lowercase",
     description = "These commands are supported:"
 )]
 pub enum Command {
-    #[command(description = "display this text.")]
+    #[command(description = "display this text")]
     Help,
-    #[command(description = "cancel the purchase procedure.")]
+    #[command(description = "cancel the purchase procedure")]
     Cancel,
     #[command(description = "create new hometask")]
     Add,
@@ -26,7 +24,16 @@ pub enum State {
     Start,
     ReceiveProductChoice,
     CreateCategorie,
-    AddTaskName { categorie: String },
-    AddDescription { categorie: String, taskname: String },
-    CreateTask { categorie: String, taskname: String, description: String },
+    AddTaskName {
+        category: String,
+    },
+    AddDescription {
+        category: String,
+        task_name: String,
+    },
+    CreateTask {
+        category: String,
+        task_name: String,
+        description: String,
+    },
 }
